Build standalone shared library for use in FFI (Closes: #152811)
Upstream provides shared-library support purely as a convenience
for the rare user who requires some foreign funion interface with
SWI-Prolog. Indeed, using swi-prolog with shared library seriously
degrades performance on some platforms.

Because of this, no care is taken to maintain ABI compatibility and
the soname is linked with the upstream version number. We install the
shared object directly in the swi-prolog package, as it has no benefit
to split it as package that links against it will have to require the
package anyway for the Standard Library.

Signed-off-by: Chris Lamb <chris@chris-lamb.co.uk>